+++ src/PhpSpreadsheet/Reader/BaseReader.php
@@ -94,7 +94,7 @@ abstract class BaseReader implements IReader
         return $this->readDataOnly;
     }
 
-    public function setReadDataOnly(bool $readCellValuesOnly): self
+    public function setReadDataOnly(bool $readCellValuesOnly): static
     {
         $this->readDataOnly = $readCellValuesOnly;
 
@@ -106,7 +106,7 @@ abstract class BaseReader implements IReader
         return $this->readEmptyCells;
     }
 
-    public function setReadEmptyCells(bool $readEmptyCells): self
+    public function setReadEmptyCells(bool $readEmptyCells): static
     {
         $this->readEmptyCells = $readEmptyCells;
 
@@ -118,7 +118,7 @@ abstract class BaseReader implements IReader
         return $this->ignoreRowsWithNoCells;
     }
 
-    public function setIgnoreRowsWithNoCells(bool $ignoreRowsWithNoCells): self
+    public function setIgnoreRowsWithNoCells(bool $ignoreRowsWithNoCells): static
     {
         $this->ignoreRowsWithNoCells = $ignoreRowsWithNoCells;
 
@@ -130,7 +130,7 @@ abstract class BaseReader implements IReader
         return $this->includeCharts;
     }
 
-    public function setIncludeCharts(bool $includeCharts): self
+    public function setIncludeCharts(bool $includeCharts): static
     {
         $this->includeCharts = $includeCharts;
 
@@ -142,7 +142,7 @@ abstract class BaseReader implements IReader
         return $this->enableDrawingPassThrough;
     }
 
-    public function setEnableDrawingPassThrough(bool $enableDrawingPassThrough): self
+    public function setEnableDrawingPassThrough(bool $enableDrawingPassThrough): static
     {
         $this->enableDrawingPassThrough = $enableDrawingPassThrough;
 
@@ -156,7 +156,7 @@ abstract class BaseReader implements IReader
     }
 
     /** @param null|string|string[] $sheetList */
-    public function setLoadSheetsOnly(string|array|null $sheetList): self
+    public function setLoadSheetsOnly(string|array|null $sheetList): static
     {
         if ($sheetList === null) {
             return $this->setLoadAllSheets();
@@ -167,7 +167,7 @@ abstract class BaseReader implements IReader
         return $this;
     }
 
-    public function setLoadAllSheets(): self
+    public function setLoadAllSheets(): static
     {
         $this->loadSheetsOnly = null;
 
@@ -179,7 +179,7 @@ abstract class BaseReader implements IReader
         return $this->readFilter;
     }
 
-    public function setReadFilter(IReadFilter $readFilter): self
+    public function setReadFilter(IReadFilter $readFilter): static
     {
         $this->readFilter = $readFilter;
 
@@ -192,7 +192,7 @@ abstract class BaseReader implements IReader
      * these can be specified within a spreadsheet
      * in a way that can subject the caller to security exploits.
      */
-    public function setAllowExternalImages(bool $allowExternalImages): self
+    public function setAllowExternalImages(bool $allowExternalImages): static
     {
         $this->allowExternalImages = $allowExternalImages;
 
@@ -213,7 +213,7 @@ abstract class BaseReader implements IReader
      *
      * @param Closure(string):bool $isWhitelisted
      */
-    public function setIsWhitelisted(Closure $isWhitelisted): self
+    public function setIsWhitelisted(Closure $isWhitelisted): static
     {
         $this->isWhitelisted = $isWhitelisted;
 
@@ -224,7 +224,7 @@ abstract class BaseReader implements IReader
      * Create a blank sheet if none are read,
      * possibly due to a typo when using LoadSheetsOnly.
      */
-    public function setCreateBlankSheetIfNoneRead(bool $createBlankSheetIfNoneRead): self
+    public function setCreateBlankSheetIfNoneRead(bool $createBlankSheetIfNoneRead): static
     {
         $this->createBlankSheetIfNoneRead = $createBlankSheetIfNoneRead;
 
@@ -346,7 +346,7 @@ abstract class BaseReader implements IReader
         return $this->valueBinder;
     }
 
-    public function setValueBinder(?IValueBinder $valueBinder): self
+    public function setValueBinder(?IValueBinder $valueBinder): static
     {
         $this->valueBinder = $valueBinder;