+++ src/PhpSpreadsheet/Reader/Csv.php
@@ -89,17 +89,17 @@ class Csv extends BaseReader
 
     protected bool $preserveNumericFormatting = false;
 
-    private bool $preserveNullString = false;
+    protected bool $preserveNullString = false;
 
-    private bool $sheetNameIsFileName = false;
+    protected bool $sheetNameIsFileName = false;
 
-    private string $getTrue = 'true';
+    protected ?string $getTrue = null;
 
-    private string $getFalse = 'false';
+    protected ?string $getFalse = null;
 
-    private string $thousandsSeparator = ',';
+    protected ?string $thousandsSeparator = null;
 
-    private string $decimalSeparator = '.';
+    protected ?string $decimalSeparator = null;
 
     /**
      * Create a new CSV Reader instance.
@@ -301,16 +301,25 @@ class Csv extends BaseReader
         }
         $this->openFile($filename);
         if ($this->inputEncoding !== 'UTF-8') {
-            fclose($this->fileHandle);
-            $entireFile = file_get_contents($filename);
-            $fileHandle = fopen('php://memory', 'r+b');
-            if ($fileHandle !== false && $entireFile !== false) {
-                $this->fileHandle = $fileHandle;
-                $data = StringHelper::convertEncoding($entireFile, 'UTF-8', $this->inputEncoding);
-                fwrite($this->fileHandle, $data);
-                $this->skipBOM();
-            }
+            $this->convertNonUtf8($filename);
         }
+    }
+
+    protected function convertNonUtf8(string $filename): void
+    {
+        fclose($this->fileHandle);
+        $entireFile = file_get_contents($filename);
+        if ($entireFile === false) {
+            throw new ReaderException("Unable to get contents of $filename"); // @codeCoverageIgnore
+        }
+        $fileHandle = fopen('php://memory', 'r+b');
+        if ($fileHandle === false) {
+            throw new ReaderException('Unable to open php://memory'); // @codeCoverageIgnore
+        }
+        $this->fileHandle = $fileHandle;
+        $data = StringHelper::convertEncoding($entireFile, 'UTF-8', $this->inputEncoding);
+        fwrite($this->fileHandle, $data);
+        $this->skipBOM();
     }
 
     public function setTestAutoDetect(bool $value): self
@@ -418,10 +427,10 @@ class Csv extends BaseReader
         $rowData = self::getCsv($fileHandle, 0, $delimiter, $this->enclosure, $this->escapeCharacter);
         $valueBinder = $this->valueBinder ?? Cell::getValueBinder();
         $preserveBooleanString = method_exists($valueBinder, 'getBooleanConversion') && $valueBinder->getBooleanConversion();
-        $this->getTrue = Calculation::getTRUE();
-        $this->getFalse = Calculation::getFALSE();
-        $this->thousandsSeparator = StringHelper::getThousandsSeparator();
-        $this->decimalSeparator = StringHelper::getDecimalSeparator();
+        $this->getTrue ??= Calculation::getTRUE();
+        $this->getFalse ??= Calculation::getFALSE();
+        $this->thousandsSeparator ??= StringHelper::getThousandsSeparator();
+        $this->decimalSeparator ??= StringHelper::getDecimalSeparator();
         while (is_array($rowData)) {
             $noOutputYet = true;
             $columnLetter = 'A';
@@ -466,9 +475,9 @@ class Csv extends BaseReader
     private function convertBoolean(mixed &$rowDatum): void
     {
         if (is_string($rowDatum)) {
-            if (strcasecmp($this->getTrue, $rowDatum) === 0 || strcasecmp('true', $rowDatum) === 0) {
+            if (strcasecmp((string) $this->getTrue, $rowDatum) === 0 || strcasecmp('true', $rowDatum) === 0) {
                 $rowDatum = true;
-            } elseif (strcasecmp($this->getFalse, $rowDatum) === 0 || strcasecmp('false', $rowDatum) === 0) {
+            } elseif (strcasecmp((string) $this->getFalse, $rowDatum) === 0 || strcasecmp('false', $rowDatum) === 0) {
                 $rowDatum = false;
             }
         } else {
@@ -484,15 +493,15 @@ class Csv extends BaseReader
         $numberFormatMask = '';
         if ($this->castFormattedNumberToNumeric === true && is_string($rowDatum)) {
             $numeric = str_replace(
-                [$this->thousandsSeparator, $this->decimalSeparator],
+                [(string) $this->thousandsSeparator, (string) $this->decimalSeparator],
                 ['', '.'],
                 $rowDatum
             );
 
             if (is_numeric($numeric)) {
-                $decimalPos = strpos($rowDatum, $this->decimalSeparator);
+                $decimalPos = strpos($rowDatum, (string) $this->decimalSeparator);
                 if ($this->preserveNumericFormatting === true) {
-                    $numberFormatMask = (str_contains($rowDatum, $this->thousandsSeparator))
+                    $numberFormatMask = (str_contains($rowDatum, (string) $this->thousandsSeparator))
                         ? '#,##0' : '0';
                     if ($decimalPos !== false) {
                         $decimals = strlen($rowDatum) - $decimalPos - 1;
@@ -759,5 +768,33 @@ class Csv extends BaseReader
     private static function getDefaultEscapeCharacter(int $version = PHP_VERSION_ID): string
     {
         return $version < 90000 ? '\\' : '';
+    }
+
+    public function setGetTrue(?string $getTrue): self
+    {
+        $this->getTrue = $getTrue;
+
+        return $this;
+    }
+
+    public function setGetFalse(?string $getFalse): self
+    {
+        $this->getFalse = $getFalse;
+
+        return $this;
+    }
+
+    public function setDecimalSeparator(?string $decimalSeparator): self
+    {
+        $this->decimalSeparator = $decimalSeparator;
+
+        return $this;
+    }
+
+    public function setThousandsSeparator(?string $thousandsSeparator): self
+    {
+        $this->thousandsSeparator = $thousandsSeparator;
+
+        return $this;
     }
 }